Rehome the base contract tests and tighten the migrated comments
